requirements.txt
camel-ai[all]
streamlit